Revert "refactor two candidate generators to stick to immutable annonation"
This reverts commit 8d2be343fbd8fbfc676bc4bfcd76ee2e5c7c8cf0.
